% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedR.R
\name{cedR}
\alias{cedR}
\title{Perform One-Way Statistical Analysis and Plotting}
\usage{
cedR(df, dependent_var, factor_var, df_file_name, logFile)
}
\arguments{
\item{df}{A data frame containing the data to analyze in a long format.}

\item{dependent_var}{Written in "", the name of the dependent variable as a string.}

\item{factor_var}{Written in "", the name of the factor variable as a string.}

\item{df_file_name}{Written in "", a string specifying the output file name.}

\item{logFile}{Written in "", a string specifying the log file name.}
}
\value{
A list containing the summary data frame and the plot object.
}
\description{
This function performs statistical analysis and creates a plot.
}
